requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.3
python-multipart==0.0.9
pydantic==2.8.2
PyPDF2==3.0.1
python-dotenv==1.0.1
httpx==0.27.0
groq==0.12.0
fastapi
uvicorn[standard]
python-dotenv
groq
pydantic
python-multipart
passlib[bcrypt]
python-jose[cryptography]
PyMuPDF
